move check_dim() to sct.image
